Cleanup (#15)
* first pass

* Fix failing test examples after mr.kruczek cleanup.

* ODA-37: adapt to lower gem versions (#16)

* ODA-37: adapt to lower gem versions

* ODA-37: clear structure and improve gemspec

* ODA-37: matrix workflow initial

* ODA-37: rubocop + less versions + no parallelism

* ODA-37: fix workflow

* ODA-37: parallel run try

* ODA-37: try to make it simple

* ODA-37: remove .ruby-version file

* ODA-37: implement version change for real

* ODA-37: ruby version reunion

* ODA-37: fix schema load

* ODA-37: check

* ODA-37: nonono

* ODA-37: rubocop

* ODA-37: try to fix rubocop

* ODA-37: clean up

* ODA-36: Implement :from config property. Add new model and change the… (#17)

* ODA-36: Implement :from config property. Add new model and change the way orderable context is injected in specs.

* ODA-36: Implement direction property - initial attempt.

* ODA-36: Normalize options before initializing config file. Delegate getters to config instead of unpacking the whole config object inside executor.

* Add transaction and review comments.

* ODA-36: Extract intemental and decremental executor to reduce number of conditions.

* ODA-36: Add test examples for model with many scopes having descending option.

* ODA-36: Add order scope direction to config and update default scope.

* chnage sort style and add ruby 2.5 tests

* dont use where.not which changed logic in AR 6.1

* Update readme.

* Implement suggestions after code review.

* define get method in executor module and simplify model extension.

* it seems that reload is unnecessary

* Add more specific test example covering destroy_all.

* Remove transaction while moving a record from scope to scope. Cover it with additional test example.

* ODA: Implement suggestions after CR. (#18)

* ODA: Implement suggestions after CR.

* ODA: Adapt code to make rubocop cops pass successfully

* ODA: Fix reset function.

* ODA: logic simplification and add reorder tests

* ODA: after CR improvements

* ODA: typo

---------


---------

+++ spec/spec_helper.rb
@@ -7,8 +7,8 @@ require "database_cleaner/active_record"
 require "shoulda-matchers"
 require "factory_bot_rails"
 require "ammeter/init"
+require_relative "./factories"
 Dir["./support/*.rb"].sort.each { |file| require file }
-Dir["./factories/*.rb"].sort.each { |file| require file }
 
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure
@@ -25,16 +25,18 @@ RSpec.configure do |config|
     FactoryBot.reload
   end
 
+  unless ENV["SPEC_DISABLE_DROP_DATABASE"] == "1"
+    config.after(:suite) do
+      Rake::Task["db:drop"].invoke
+    end
+  end
+
   config.before(:suite) do
     Rake::Task["db:create"].invoke
     Rake::Task["db:migrate"].invoke
 
     DatabaseCleaner.strategy = :transaction
     DatabaseCleaner.clean_with(:truncation)
-  end
-
-  config.after(:suite) do
-    Rake::Task["db:drop"].invoke
   end
 
   config.around(:each) do |example|
